Add block skipping logic to block archive repos instead of the base repo
